+++ Platform/settings/production.py
@@ -17,6 +17,10 @@ ALLOWED_HOSTS = [DOMAIN]
 
 # Make sure SECRET_KEY is set to a secret in production
 SECRET_KEY = os.environ.get("SECRET_KEY", None)
+# Make sure IDENTITY_RSA_PRIVATE_KEY is set to a secret in production
+IDENTITY_RSA_PRIVATE_KEY = os.environ.get("IDENTITY_RSA_PRIVATE_KEY", None)
+if IDENTITY_RSA_PRIVATE_KEY is None:
+    raise Exception("Please provide environment variable IDENTITY_RSA_PRIVATE_KEY in production")
 
 # Sentry settings
 SENTRY_URL = os.environ.get("SENTRY_URL", "")
